doc: add more documentation around SP 800-131A
At times it is desired to compile out dissallowed, deprecated, or
legacy behaviour out of the FIPS provider. This might be done due to
desire to protect information post the future transition dates, or to
achieve full length FIPS certification.

This document is a living document to keep track of the revisions to
SP 800-131A as well as any configuration gaps.

Currently revision 2 is documented together with known compile-time
toggles. The revision 3 draft will be documented once outstanding gaps
are addressed.